Refactor Menu to isolated component
